requirements.txt
fastapi==0.87.0
uvicorn==0.19.0
sqlalchemy==1.4.44
mysql-connector-python==8.0.31
pydantic~=1.10.2
PyMySQL~=1.0.2
cryptography==38.0.3
starlette==0.22.0
python-multipart==0.0.5
Pillow==9.3.0